% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-allcategory.R
\docType{data}
\name{allCategory}
\alias{allCategory}
\title{All possible values from the CATEGORY field.}
\format{
A vector with 1 variable
}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allCategory
}
\description{
All possible values from the CATEGORY field.
}
\keyword{datasets}
