% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-allcommodity.R
\docType{data}
\name{allCommodity}
\alias{allCommodity}
\title{All possible values from the COMMODITY field.}
\format{
A vector with 1 variable
}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allCommodity
}
\description{
All possible values from the COMMODITY field.
}
\keyword{datasets}
