% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-allcounty.R
\docType{data}
\name{allCounty}
\alias{allCounty}
\title{All possible values from the COUNTY field.}
\format{
A vector with 1 variable
}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allCounty
}
\description{
All possible values from the COUNTY field.
}
\keyword{datasets}
