% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-alldataitem.R
\docType{data}
\name{allDataItem}
\alias{allDataItem}
\title{All possible values from the DATA ITEM field.}
\format{A vector with 1 variable}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allDataItem
}
\description{
All possible values from the DATA ITEM field.
}
\keyword{datasets}
