% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-alldomain.R
\docType{data}
\name{allDomain}
\alias{allDomain}
\title{All possible values from the DOMAIN field.}
\format{A vector with 1 variable}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allDomain
}
\description{
All possible values from the DOMAIN field.
}
\keyword{datasets}
