% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-allgeoglevel.R
\docType{data}
\name{allGeogLevel}
\alias{allGeogLevel}
\title{All possible values from the GEOGRAPHY LEVEL field.}
\format{
A vector with 1 variable
}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allGeogLevel
}
\description{
All possible values from the GEOGRAPHY LEVEL field.
}
\keyword{datasets}
