% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-allgroup.R
\docType{data}
\name{allGroup}
\alias{allGroup}
\title{All possible values from the GROUP field.}
\format{A vector with 1 variable}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allGroup
}
\description{
All possible values from the GROUP field.
}
\keyword{datasets}
