% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-allprogram.R
\docType{data}
\name{allProgram}
\alias{allProgram}
\title{All possible values from the PROGRAM field.}
\format{A vector with 1 variable}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allProgram
}
\description{
All possible values from the PROGRAM field.
}
\keyword{datasets}
