% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-allsector.R
\docType{data}
\name{allSector}
\alias{allSector}
\title{All possible values from the SECTOR field.}
\format{A vector with 1 variable}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allSector
}
\description{
All possible values from the SECTOR field.
}
\keyword{datasets}
