% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-allstate.R
\docType{data}
\name{allState}
\alias{allState}
\title{All possible values from the STATE field.}
\format{
A vector with 1 variable
}
\source{
\url{https://quickstats.nass.usda.gov}
}
\usage{
allState
}
\description{
All possible values from the STATE field.
}
\keyword{datasets}
