% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllCommodity}
\alias{getAllCommodity}
\title{Get all possible values from COMMODITY field}
\usage{
getAllCommodity(key)
}
\arguments{
\item{key}{your USDA api key}
}
\value{
All possible values of PROGRAM
}
\description{
Get all possible values from COMMODITY field
}
\examples{
\dontrun{getAllProgram(key)}
}
