% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllCounty}
\alias{getAllCounty}
\title{Get all possible values from COUNTY field}
\usage{
getAllCounty(key)
}
\arguments{
\item{key}{your USDA api key}
}
\value{
All possible values of PROGRAM
}
\description{
Get all possible values from COUNTY field
}
\examples{
\dontrun{getAllProgram(key)}
}
