% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllDataItem}
\alias{getAllDataItem}
\title{Get all possible values from DATA ITEM field}
\usage{
getAllDataItem(key)
}
\arguments{
\item{key:}{your USDA api key}
}
\description{
Get all possible values from DATA ITEM field
}
\examples{
x <- tidyUSDA::getAllDataItem(mykey)

 # View the vector in RSTudio in order to search for terms 
 View(x)
}
