% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllGeogLevel}
\alias{getAllGeogLevel}
\title{Get all possible values from GEOGRAPHIC LEVEL field}
\usage{
getAllGeogLevel(key)
}
\arguments{
\item{key:}{your USDA api key}
}
\description{
Get all possible values from GEOGRAPHIC LEVEL field
}
\examples{
x <- tidyUSDA::getAllGeogLevel(mykey)

 # View the vector in RSTudio in order to search for terms 
 View(x)
}
