% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllGroup}
\alias{getAllGroup}
\title{Get all possible values from GROUP field}
\usage{
getAllGroup(key)
}
\arguments{
\item{key}{your USDA api key}
}
\value{
All possible values of PROGRAM
}
\description{
Get all possible values from GROUP field
}
\examples{
\dontrun{getAllProgram(key)}
}
