% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllProgram}
\alias{getAllProgram}
\title{Get all possible values from PROGRAM field}
\usage{
getAllProgram(key)
}
\arguments{
\item{key:}{your USDA api key}
}
\description{
Get all possible values from PROGRAM field
}
\examples{
x <- tidyUSDA::getAllProgram(mykey)

# View the vector in RSTudio in order to search for terms 
View(x)
}
