% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllSector}
\alias{getAllSector}
\title{Get all possible values from SECTOR field}
\usage{
getAllSector(key)
}
\arguments{
\item{key:}{your USDA api key}
}
\description{
Get all possible values from SECTOR field
}
\examples{
x <- tidyUSDA::getAllSector(mykey)

 # View the vector in RSTudio in order to search for terms 
 View(x)
 
}
