% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllState}
\alias{getAllState}
\title{Get all possible values from STATE field}
\usage{
getAllState(key)
}
\arguments{
\item{key}{your USDA api key}
}
\value{
All possible values of PROGRAM
}
\description{
Get all possible values from STATE field
}
\examples{
\dontrun{getAllProgram(key)}
}
