% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuickstat.R
\name{getQuickstat}
\alias{getQuickstat}
\title{getQuickstat}
\usage{
getQuickstat(key = NULL, program = NULL, data_item = NULL,
  sector = NULL, group = NULL, commodity = NULL, category = NULL,
  domain = NULL, geographic_level = NULL, state = NULL,
  county = NULL, year = NULL, geometry = FALSE, lower48 = FALSE)
}
\arguments{
\item{key}{your USDA api key. Get one at https://quickstats.nass.usda.gov/api}

\item{program}{program field}

\item{data_item}{data_item field}

\item{sector}{sector field}

\item{group}{group field}

\item{commodity}{commodity field}

\item{category}{category field}

\item{domain}{domain field}

\item{geographic_level}{geographic_level field}

\item{state}{state field}

\item{county}{county field}

\item{year}{year field}

\item{geometry}{geometry field (TRUE or FALSE), set to TRUE if you would like a simple features (SF) geometry field included. 
Only works when geographic_level is set to 'COUNTY' or 'STATE'}

\item{lower48}{limit data to the lower 48 states?}
}
\description{
Get values from Quickstats in a dataframe or data frame
with sf:: geometry field
}
\note{
Go to the webpage https://quickstats.nass.usda.gov/. As a best practice, select the items in these fields and test that that data item 
   exists in the browser before using those parameters in this function. When you have a dataset that works, enter those values in the 
   function as parameters. Ideally, only enter values for your key obviously, then PROGRAM, DATA_ITEM, GEOGRAPHIC_LEVEL and then if
   necessary, DOMAIN, STATE, COUNTY or YEAR.
}
