% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyUSDA.R
\docType{package}
\name{tidyUSDA}
\alias{tidyUSDA}
\alias{tidyUSDA-package}
\title{tidyUSDA: An Interface to USDA QuickStats Data with Mapping Capabilities.}
\description{
Lets you query the USDA QuickStats data from R, and join in geometries for plotting.
}
