% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllCategory}
\alias{getAllCategory}
\title{Get all possible values from CATEGORY field}
\usage{
getAllCategory(key)
}
\arguments{
\item{key}{your USDA api key}
}
\value{
All possible values of PROGRAM
}
\description{
Get all possible values from CATEGORY field
}
\examples{
\dontrun{getAllProgram(key)}
}
