% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllDomain}
\alias{getAllDomain}
\title{Get all possible values from DOMAIN field}
\usage{
getAllDomain(key)
}
\arguments{
\item{key}{your USDA api key}
}
\value{
All possible values of PROGRAM
}
\description{
Get all possible values from DOMAIN field
}
\examples{
\dontrun{getAllProgram(key)}
}
