% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllGroup}
\alias{getAllGroup}
\title{Get all possible values from GROUP field}
\usage{
getAllGroup(key)
}
\arguments{
\item{key:}{your USDA api key}
}
\description{
Get all possible values from GROUP field
}
\examples{
x <- tidyUSDA::getAllGroup(mykey)

 # View the vector in RSTudio in order to search for terms 
 View(x)
}
