% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllProgram}
\alias{getAllProgram}
\title{Get all possible values from PROGRAM field}
\usage{
getAllProgram(key)
}
\arguments{
\item{key}{your USDA api key}
}
\value{
All possible values of PROGRAM
}
\description{
Get all possible values from PROGRAM field
}
\examples{
\dontrun{getAllProgram(key)}
}
