% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllSector}
\alias{getAllSector}
\title{Get all possible values from SECTOR field}
\usage{
getAllSector(key)
}
\arguments{
\item{key}{your USDA api key}
}
\value{
All possible values of PROGRAM
}
\description{
Get all possible values from SECTOR field
}
\examples{
\dontrun{getAllProgram(key)}
}
