% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getAllState}
\alias{getAllState}
\title{Get all possible values from STATE field}
\usage{
getAllState(key)
}
\arguments{
\item{key:}{your USDA api key}
}
\description{
Get all possible values from STATE field
}
\examples{
x <- tidyUSDA::getAllState(mykey)

 # View the vector in RSTudio in order to search for terms 
 View(x)
}
