% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotUSDA}
\alias{plotUSDA}
\title{plotUSDA}
\usage{
plotUSDA(df)
}
\arguments{
\item{df}{a dataframe with a simple feature column}
}
\description{
Quickly plot a dataframe produced by the getQuickstat() function.
}
\examples{
\dontrun{
# Use output from getQuickstat()
plotUSDA(df = df_from_getQuickstat)
}

}
