% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotUSDA}
\alias{plotUSDA}
\title{plotUSDA}
\usage{
plotUSDA(df, title = "USDA Plot")
}
\arguments{
\item{df}{a dataframe with a simple feature column}

\item{title}{title for the plot}
}
\description{
Quickly plot a dataframe produced by the getQuickstat() function.
}
