% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyUSDA.R
\docType{package}
\name{tidyUSDA}
\alias{tidyUSDA}
\alias{tidyUSDA-package}
\title{tidyUSDA: An interface to USDA Quick Stats data with mapping capabilities.}
\description{
The package lets you query the USDA QuickStats data from R, and join in geometries.
}
