% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyUSDA.R
\docType{package}
\name{tidyUSDA}
\alias{tidyUSDA}
\title{tidyUSDA: An Interface to USDA QuickStats Data with Mapping Capabilities.}
\description{
A minimal toolset for gathering USDA Quick Stat data for analysis and visualization.
}
