% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyUSDA.R
\docType{package}
\name{tidyUSDA}
\alias{tidyUSDA}
\alias{tidyUSDA-package}
\title{tidyUSDA: An Interface to USDA QuickStats Data with Mapping Capabilities.}
\description{
A minimal toolset for gathering USDA Quick Stat data for analysis and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bradlindblad.github.io/tidyUSDA}
  \item \url{https://github.com/bradlindblad/tidyUSDA}
}

}
\author{
\strong{Maintainer}: Brad Lindblad \email{bradley.lindblad@gmail.com}

Other contributors:
\itemize{
  \item Michael Thomas \email{mthomas@ketchbrookanalytics.com} [contributor]
  \item Alex Mindeman \email{alexandramindeman@gmail.com} [contributor]
}

}
